Move the Config class to a dedicated file
